medfastai3/server/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-socketio
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
